% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dismod_to_stgpr.R
\name{dismod_to_stgpr}
\alias{dismod_to_stgpr}
\title{DisMod to ST-GPR Bundle Converter}
\usage{
dismod_to_stgpr(input_dismod_bundle)
}
\arguments{
\item{input_dismod_bundle}{Bundle data in DisMod shape.}
}
\value{
The inputted bundle data, converted to ST-GPR shape.
}
\description{
DisMod to ST-GPR Bundle Converter
}
