% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stgpr_to_dismod.R
\name{stgpr_to_dismod}
\alias{stgpr_to_dismod}
\title{ST-GPR to DisMod Bundle Converter}
\usage{
stgpr_to_dismod(input_stgpr_bundle)
}
\arguments{
\item{input_stgpr_bundle}{A crosswalk version of a ST-GPR shape bundle.}
}
\value{
The inputted bundle data, converted to DisMod shape.
}
\description{
ST-GPR to DisMod Bundle Converter
}
\examples{
stgpr_to_dismod(diabetes_stgpr_bundle)
}
