% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_means.R
\name{weighted_mean}
\alias{weighted_mean}
\title{Weighted Means Function to Age-Standardize GBD Outputs}
\usage{
weighted_mean(
  df,
  weight_by,
  gbd_round_id = 7,
  decomp_step = "iterative",
  is_model_results = TRUE
)
}
\arguments{
\item{df}{A data frame produced using the shared function "get_model_results"}

\item{weight_by}{A character string of variables to weight by}

\item{gbd_round_id}{The gbd_round_id for the get_location_metadata shared function, the default value is 7}

\item{decomp_step}{The decomp_step for the get_population and get_location_metadata shared functions, the default value is "iterative"}

\item{is_model_results}{A boolean variable, set as 'TRUE' (the default) if the input is model_results; if 'FALSE', the user must enter a location_set_id}
}
\value{
A data frame with age-standardized data.
}
\description{
Weighted Means Function to Age-Standardize GBD Outputs
}
